+++ include/utils.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+typedef struct {
+    float r;
+    float g;
+    float b;
+}RGB;
+
+RGB hex2rgb(int hexValue);
+
+typedef struct {
+    float x;
+    float y;
+}Pos;
+
+
